Semantics of `open` is not `include`.
The desugar modules patch inadvertently changed the semantics of
`open` to be that of `include`. This regression was spotted by
@SimonJF in #606. From reading the generated code I mistakenly
concluded that it was a loader problem. However, as pointed out by
@SimonJF, the chaser is supposed to lift dependencies into "synthetic"
modules in the combined program.

With that in mind, it became clear to that the desugar modules patch
inadvertently changed the semantics of `open`, because I heedlessly
added the bindings resulting from an `open` to the module scope that
was being built. As a result they would correctly shadow the current
bindings in scope, but they would also inadvertently become part of
the module.

This patch rectifies this mistake. The implementation still makes use
of two-level scope data structure, however, now it only keeps track of
the current scope and its "delta", where the delta consists of all the
bindings defined in the current immediate scope. Bindings resulting
from an `open` are not added to the delta, and thus, do not become
part of the module's implicit signature.

Fixes #606.