Improve the REPL efficacy when using modules.
The REPL directive `@load` loads a source file and pulls its contents
into the current scope. If the source file defines any modules the
user cannot access them following a load, because the state of module
desugarer is not preserved between `@load` directives when in
interactive mode.

This patch rectifies this issue such that the state is preserved
between subsequent invocations of the module desugarer. The state is
only preserved when Links is operating is interactive mode.