Disallow alien blocks below top-level.
This patch provides an easy fix for #604 by disallowing alien blocks
syntactically below the top-level. Incidentally this makes the alien
and alien blocks more coherent as it is already the case that the
former cannot appear below the top-level.

Note that alien and alien blocks can still appear below the global
top-level if they appear in a local module. However, they are still
syntactically at the "top-level" of the said module.

Fixes #604.