[Handlers] Generalise parameterised handlers.
This patch generalises the kind of expression which is allowed to
appear in the expression position of a parameter binding form in a
parameterised handler. It was necessary to slightly change the syntax
of parameter binding to allow every expression form. Previously the
binding form syntax was

      logical_expression -> pattern

now it is

      pattern <- expression

The syntactic swap between the expression and pattern was necessary to
avoid a parser conflict.

The patch also rectifies a bug with shallow handlers which meant the
empty clause set was not allowed.