Lexer hack to enable use of chevrons in effect patterns.
This commit augments the lexer context such that chevrons (`<` and
`>`) can be used as syntax to denote an operation pattern. The lexer
context is augmented with an additional mode called `effect_mode`
which is intended to be `true` whenever an operation pattern is
allowed to occur, and `false` whenever a XML literal is allowed to
occur. The following rules decide in which mode the lexer operates:

- if the current lexeme is `case` and `effect_mode = false`, then set
  `effect_mode = true` and return `CASE`.

- if the current lexeme is `<` and `effect_mode = true`, then increment
  the number of seen `<` and return `LANGLE`. Otherwise apply the rule
  for operators.

- if the current lexeme is `>` and `effect_mode = true`, then decrement
  the number of seen `<` and return `RANGLE`. Otherwise apply the rule
  for operators.

- if the current lexeme is `<def_qname`, `effect_mode = true`, and
  `def_qname` is a valid constructor name, then increment the number
  of seen `<`, push `CONSTRUCTOR def_qname` to the token queue, and
  return `LANGLE`. Otherwise apply the rule for XML start tags.

- if the current lexeme is `->`, `effect_mode = true`, and the number
  of seen `<` is 0, then set `effect_mode = false` and return
  `RARROW`. Otherwise return `RARROW`.

Scanner inspects the lexer context before invoking the next lexer. If
the token queue is nonempty, then the head token is popped and
returned, otherwise the next lexer is applied with `lexbuf`.