Abstract binders (#646)
* Abstract binders

This patch makes the frontend representation of binders abstract. It
turns out that an abstract representation is convenient for enabling
the proper module system (#603) to live along side the current module
system as the former benefits from a slightly richer notion of
binders.

* Update the given binder during AST traversals.

Since binders are abstract we should not create fresh binders during
AST traversals. Instead we should update the given binder using the
`Binder` and `WithPos` interfaces.