Removes the `dumpTypes` function from the Links standard library.
This patch removes the built-in function `dumpTypes`. Judging from the
comment in `dumpTypes.ml` it seems that its sole purpose was to
provide some interface for displaying type information in an IDE. The
comment even states that it is a hack. Hack or not, the function does
not work any longer as it triggers an internal failure on use. My
guess is that this has been the case for a while.

I would like to remove this function as it specifically requires
access to the `prelude` typing environment. Access was provided via a
global hack which is at odds with how I would like to structure the
new modular loader (c.f. #603).