Add the `Mono' restriction
This requires that any types which unify with it are monomorphic, thus
ensuring cohesion of FreezeML.

When visiting a function body, we attempt to make all arguments Mono,
skipping patterns which have an explicit type annotation. This is simply
done by calling Types.mono.make_type, which will convert all compatible
flexible type variables, and ignore everything else. This allows us to
only touch patterns without an explicit type annotation.

When generalising, we convert any remaining Mono quantifiers into Any
ones.

Note, there is one bug with the current implementation, in which Mono
variables may be converted into Session ones, which may then be
polymorphic. The obvious solution would be to make monomorphicity (?) a
separate property of the subkind, but doing so seems unnecessarily
complex at this stage.