Make the session examples a bit closer to bad1/bad2
This is as close as I've been able to get - anything closer requires us
to bind f to a variable, which in turn forces it to have the Mono
restriction.

I both these examples, the initial `s:%::Session' on the function binder
means the Mono restriction is not applied, and so none of our usages are
constrained.