Fix Mono -> Any weakening touching the outer type variables
When we used generalise_with_subst, we converted the outer type
variables to Any, rather than the inner quantifiers/type variables,
causing the subkinds of terms to change.

This is going to start causing a lot of tests to fail, as they'll all
now have the `Mono` restriction appearing in their types. This is
technically correct, in the sense that this is the originally inferred
type, but probably not desirable - probably worth discussing how/if
typeSugar should be changed.