Add sugar for type applications with an implicit effect variable
Consider the type

    typename Comp(e::Eff, a) = () -e-> a

With this extension, one may write functions such as `Comp(a) -a-> ()',
and this will be converted to `Comp({ |e}, a) -e-> ()' while desugaring.

This hopefully allows us to write types involving effects with a little
less syntactic overhead.

Note, the printer does not currently handle such syntactic sugar.