Record field punning (#667)
* Record field punning

A feature I've been missing from OCaml is record field punning. In
OCaml, we can write:

```
let hello = "hello!" in
{ hello; world = 5 }
```

instead of

```
let hello = "hello!" in
{ hello = hello; world = 5 }
```

This patch is a simple parser hack to allow the same sort of thing in
Links. We can't do field punning directly since we get a reduce/reduce
with a parenthesised expression, but we can delineate punned fields with
an operator (this patch uses `=`), allowing us to write:

```
var hello = "hello!";
(=hello, world = 5)
```

which is parsed in as

```
var hello = "hello!";
(hello = hello, world = 5)
```

* Field punning in record deconstruction

* Change punning marker to =

* Update tests

* Add final few tests