Some minor cleanup for OCaml >=4.08 (#668)
* Rename the "development" profile to "dev"

Dune provides two profiles by - "dev" and "release". By using "dev",
instead of our own "development" one, dune commands will use this by
default, which makes everything a little more convenient.

* Do not error on build warnings in release mode

While failing fast may be somewhat appealing, it does mean changes in
dependencies (or OCaml itself) cause the Links build to fail.

It's worth noting that the default in Dune is to have stricter warnings
for development than for release, so it's probably best to follow that.

* Remove any explicit references to the prelude

Not strictly needed, but most of these were relics of when OCaml didn't
have (|>).

* Remove all remaining usages of Pervasives

* Suppress the partially applied function warning in a few places

* Remove the wrapped workaround

Not sure what Dune changed, but it appears to no longer be required.