Improve the heuristic for when to insert shared effect variables
Previously, we would insert the shared effect variable at every
arrow/type alias. For instance:
  ((a) -> b) -> ([a]) -> [b]
would become
  ((a) -e-> b) -e-> ([a]) -e-> [b]

We now only insert the effect variable on the right-most suitable
term. In the above case, this is the `(a) -> b` and `([a]) -> [b]`
arrows, meaning we now type this as
  ((a) -e-> b) -_-> ([a]) -e-> [b]

A "suitable term" is either an arrow type or type application where the
first argument is an implicit effect type.

This change allows us to use this sugar in many more places. For
instance, we can now represent functions which map from computations to
computations, or return computations without the "outer" function
sharing the effects.