Suppress generalisation on the toplevel for typed_ir tests
It appears that the IR does not currently bind type variables defined in
vars, which means that we still have free type variables.

Instead, we just add an option to enable/disable generalisation (which
may be useful anyway), and disable it for the typed IR tests.

This is a rather horrible, but works for now.