Get the prelude working with the new effect sugar
Both spawn* (from lib.ml) and the various fork methods take an effect
polymorphic variable as an argument. However, this is incorrect -
spawn { do X } will error, as X does not have a handler.

The correct thing to do here is give the spawn methods the signature
of:

    (() { SessionFail:[||], here{p} }~@ _) ~> Process({ here{p} })

However, doing so requires several changes to both type checking and
desugaring. For now, we just bodge it and patch the prelude to have the
correct effects.