Frozen lets (#688)
* Freeze recursive functions too

We do this a little later, after desugarInners. Otherwise we get all
sorts of complications with inner types not being generalised.

Is this especially elegant? No. Does it work? I guess.

* Syntax!!

(There remains an issue with quantifiers escaping their scope when mixing frozen and unfrozen functions in a mutual block, but we can fix that later.)