Display an alert if `jslib` fails to load.
This patch instruments the Links web runtime with a "readiness check"
before calling `_startRealPage`. If the readiness check fails it emits
an error message (in HTML) and discontinues the start up process. As a
consequence Links will no longer render the user-defined webpage in
the event of the runtime dependency `jslib` failing to load. Instead the following error message is displayed:

<h1>Fatal error: Runtime dependency `jslib.js' is not loaded.</h1>

Currently, the readiness check only checks whether `jslib` is
loaded. It could be extended to check whether alien JavaScript sources
have been loaded too. However, I think the page compiler is due a
refactor before we go down that path.

Resolves #590.