Fix #678 (Quantifiers escape their scopes) (#710)
This patch does a fairly simple fix which ensures that given an explicitly-quantified signature, that the quantifiers do not appear in the typing environment after typechecking the body. The check is only performed when a function has explicit quantifiers.


As an example, the program from #678:
```
var f = id(id);

sig g : forall a. (a) -> a
fun g(x) { f(x) }
```

Now gives the error:
```
squid-bug.links:4: Type error: The quantifiers in the type of function
    g: forall a.(a) -b-> a
escape their scope, as they are present in the types:
    f: (a) -b-> a
In expression: fun g(x) { f(x) }.
```