Remove the option `realpages`. (#713)
The option `realpages` used to govern whether Links would serve
web pages built on the server or build them dynamically on the
client. The option appears to have been unused (and unsupported)
for a while. Although, the default value of `realpages` was `false`
the Links appserver would always serve pages built on the server.

This patch removes the option altogether.