Decouple IR type structure simplifications from IR type checking. (#718)
* Decouple IR type structure simplifications from IR type checking.

This patch resolves #561 by separating the type structure
simplification pass from the IR type checker. A new flag
`simplify_types` is introduced which dictates whether the
simplification pass is run -- because the IR type checker depends on
the `ElimBodiesFromMetaVars` pass the simplification pass will also,
as a temporary workaround, run if `typecheck_ir` is set to `true`.

The patch also fixes a couple of issues with side effects in the IR
backend pipeline infrastructure. As settings `typecheck_ir` and
`show_compiled_ir_after_backend_transformations` had no effect if
Links was not booted with those settings set to true, meaning that
they could not be toggled from within the REPL. This is rectified by
replaying the side effects whenever a backend pipeline is invoked.

* Disable the alias expansion transformation.