Endow testscripts with an optional header section.
This patch aims to remove some of the "magic" in the test suite
runners. Specifically, this patch augments our ad-hoc testscript
language with an optional header section, which contains key value
pairs (currently only "config: path/to/file.config" is supported) that
can be used to fix parameters globally for the tests in the respective
script. So if a script uses a particular config file it has to be
listed explicitly in the script header.

Syntatically a header is surrounded by a pair of `---`. A header may
only contain key-value pairs of the form `key: value`. This syntax is
adopted directly from the Markdown header format. Example:

```
---
config: tests/typed_ir.tests.config
---

Anonymous constant function
fun(x) {1}
stdout : fun : (_::(Unl,Mono)) -> Int
```

It is possible to override the value of the `config` key by explicitly
passing the name of a configuration file as a command line
argument. For example the following command

```
$ ./test-harness tests/typed_ir.tests tests/typed_ir.tests.config-alternative
```

would run `typed_ir.tests` with the alternative configuration file.