Compile session exceptions using deep handlers (#744)
There is an arbitrary choice as to whether to compile session exceptions
using deep or shallow handlers. I switched this to Shallow at some
point, and since then it seems that shallow handlers no longer compile
on the client.

(Unfortunately, this means that session exceptions are broken in 0.9).
This patch switches them back to deep.