Modular settings. (#745)
Modular settings.

Prior to the merge of effects and session #131, the effect handlers
patch #227, and the jbuilder/dune patch #254 it was possible to define
settings on a per compilation basis, except for settings which were
settable via the commandline. They had to be defined early before the
CLI arguments were parsed. Moreover, some compilation unit were
sensitive to as to when a setting was set, because they would use the
value of a setting to select an implementation of a signature at the
toplevel.

This patch restores the ability to define settings on a compilation
unit basis, and it does so for all settings -- meaning commandline
settable settings can be defined locally too.

At glance this patch contains the following changes:

* Local definition of settings.
* Small DSL for defining settings and their CLI extensions.
* No universal type. Settings are now GADTs, meaning reading and
  writing operations on settings are typed. As a consequence values
  are no longer restricted by a universal type.
* Boolean settings (flags) can have dependencies. Enabling a setting
  with dependencies will cause its dependencies to be switched on too.
* GNU convention compatible handling of CLI arguments.
* Retrieval of commandline arguments in Links programs via `getArgs`.
* Automatic gathering and printing of CLI settings.

To illustrate the latter point, passing `-h` or `--help` to Links on
the commandline causes Links to print something along the lines of

```
usage: links.exe [options] [source-files [-- arguments]]

Options are:
     --config=<file>             Initialises Links according to the given configuration file (default: $HOME/.opam/4.08.1/etc/links/config)
 -d, --debug                     Prints internal debugging information (development) (default: false)
     --enable-handlers           Enables the effect handlers extension (default: false)
 -e, --evaluate=<expression>     Evaluates an expression
 -h, --help                      Print help message and exit
     --path=<dir[,dir']...>      Search paths for Links modules (default: .)
     --measure_performance       Instruments the server-side runtime to measure various performance characteristics (default: false)
 -r, --rlwrap                    Selects whether to use the native readline support in REPL mode (default: true)
     --optimise                  Optimises the generated code (default: false)
     --print-keywords            Print keywords and exit
     --session-exceptions        Enables the session exceptions extension (default: false)
     --set=<setting=value>       Sets the value of a particular setting via the commandline
 -v, --version                   Print version and exit
 -w, --web-mode                  Start Links in web mode (default: false)
```

Resolves #247, resolves #290, resolves #537, resolves #551.