Print an error message if database driver loading fails (#748)
This patch slightly improves error messages when Links fails to load a
database driver. Beforehand a raw exception would be printed to
stderr. With this patch a standard error message, conforming with the
style of messages in `errors.ml`, is printed.

For example, before Links would print `LocateFailure "postgresql"`
when it failed to locate the `postgresql` driver, now it will print

```
***: Error: Cannot locate database driver 'postgresql'
```

Resolves #746