Code serialisation refactor
This patch simplifies the [value.ml] compilation unit by hoisting the
code serialisation logic from there into a separate compilation
unit. The primary motivation for this refactor is to eliminate the
`value -> desugarDatatypes` dependency which is a prerequisite for an
upcoming patch.

The JSON serialisation logic is somewhat a mess, and I have made no
attempt to rectify it. I will leave it as a future task.