Distinction between nominal and structural 'names'.
The module `Name` in `commonTypes.ml` is used to name both nominal and
structural things. It is conceptually wrong to mix the two. This patch
introduces a new module `Label` in `commonTypes.ml` which is used to
name structural things, reserving `Name` to be used for nominal
things. Although this change is only cosmetic, it becomes important as
we move towards hygienic (nominal) name handling in the frontend.