Iterative resolveServerValue in jslib.js (#752)
Before, resolveServerValue was a non-tail-recursive function. This was
problematic for larger JSON values.

Since the ordering of the resolution does not seem to matter, this patch
changes the function to an iterative queue-based algorithm. This means
we can have much better scalability (and indeed this makes large initial
state dumps for my GtoPdb application work).