Structured variable usage tracking
This patch introduces a thin abstraction for variable usage tracking
in the frontend type checker. Usages are now manipulated via a
structured interface `Usage`, as a result it is nolonger necessary
(nor possible) to manipulate them on an ad-hoc basis via bare
`StringMap`s, `StringSet`s, and lists.

This code is extracted and adapted from the name hygiene patch which I
have been working on lately. I think this code is useful on its own as
it makes the type checker more robust. Ancedotally, I did find and fix
a couple of linearity bugs in the implementation of effect handlers
and session exceptions, where usage were tracked properly.

This patch adds a small compatibility module `Ident` which is
essentially an artefact of my hygienic names patch. I have added it to
make it easier for me to eventually merge my name patch.