Fix installation of `python3` on Azure Pipelines macOS image (again).
It seems there have been two changes to either `brew` or the macOS
version on Azure Pipelines lately. It seems that the macOS image now
ships with `python` version `3.7.5` which also appears to be the
latest version maintained by the `brew` repository that we are
using. Apparently the command `brew upgrade python` fails whenever the
installed `python` package is up-to-date.

Thus this patch simply removes the `brew upgrade python` step.