Foreign Function Interface clean up. (#783)
* Foreign Function Interface clean up.

This patch refactors the implementation of FFI. The changes
are only cosmetic.

This is backported from my work on name hygiene/compilation
units. Ultimately, I would like to use the FFI to bootstrap (in the
frontend) the Links builtins library (aka `lib.ml`). By doing this
`lib.ml` need not be treated specially by the frontend, which makes it
significantly easier to implement a uniform infrastructure for
compilation units.