Accept the empty program as a valid Links program. (#787)
This patch extends the grammar of Links such that the empty program is
accepted by the parser as a syntactically valid program.

It is useful to be able to start Links with an empty prelude, e.g. for
debugging. We already accept the empty expression as valid input in
interactive mode.