Always print redefined value names and typenames in the REPL.
Prior to this patch, defining two functions with the same name would
only cause the former to be printed by the REPL, e.g.

```
links> fun f(x) { x };
f = fun : (a::Any) -> a::Any
links> fun f(x) { x };
links>
```

With this patch the REPL will also print the latter entry.

The previous implementation would simply diff the previous compiler
context and the current compiler context to determine which names were
new. However, this is not sufficient to tell whether a name has been
redefined as long as the internal names are not hygienic.

This patch refines the diffing approach:

- For values the REPL computes the set of new IR vars -- which are
  globally unique -- and uses those to determine whether a name has
  been redefined, and thus should be printed.

- As for typenames, the REPL makes use of a slight hack. Typenames are
  represented as plain strings, so in order to check whether a
  typename has been redefined this implementation performs a reference
  equality check on the object that the typename is bound to. If that
  check returns false, then the name has been redefined.

This situation will improve once I manage to land the hygienic names patch.