[WIP] Begin fixing scheduler bug
This moves `value` from `evalir.ml` into the Lwt monad. This is
necessary because `Proc.atomically` (used in `ApplyPure`) was re-calling
the main LWT entry point. We got away with this before, but it's an
actual error in LWT5 (and rightly so).

Still some bugs---primarily with XML---but we respect this invariant now.