Operator associativity and precedence (#806)
Prior to this patch the associativity and precedence of user-definable
operators would be stored in the internal state of the lexer. This
state was not preserved between invocations of the lexer on per
compilation unit and per REPL entry. As a consequence, during
subsequent parses every previously defined operator would default to
be left associative and have precedence 9.

This patch removes the operator associativity and precedence
information from the lexer. Instead of handling associativity and
precedence in the lexer and parser, they are now handled by the
transformation `shunting` which applies a variation of the Shunting
Yard algorithm on expressions. The transform uses our transformation
infrastructure to preserve and pass an operator table between
subsequent invocation of this transform. As a result both the lexer
and parser are greatly simplified. I believe this patch opens the door
for more rule simplifications in the parser, but I will leave that for
future work.

Currently, a default operator table is hard-coded in the `shunting`
module. My front-end bootstrapping patch hoists this table out of the
compiler and into a Links source file.