Abstract syntax for SQL modifications (#832)
# Motivation
Previously, SQL update and deletion operations were constructed as strings in `query.ml`. Since insertions are implementation-dependent, the picture there is even worse: we needed to do some gymnastics to get a bunch of strings, which were then passed to the `Value.database` implementation.

When we're wanting to do query rewrites (say, if we wanted to do temporal database operations, for example), the string-based approach is very inflexible. Furthermore, insertions are treated non-uniformly, which is a pity.

# Changes
This patch adds abstract syntax for SQL insertions, updates, and deletions in the `Sql` module, and adapts the rest of the codebase to use them. More specifically:

  - `Sql` contains the new abstract syntax nodes `Insert`, `Update`, and `Delete`.
  - `Sql.string_of_query` takes a field quotation function of type `string -> string`, rather than taking a `Value.database`. In turn, this breaks the dependency from `Sql` to `Value`.
  - We add a `string_of_query` method to `Value.database`, which is an overridable implementation-specific query serialisation function which calls into `Sql.string_of_query` by default.
  - The `make_insert_query` method is deleted from `Value.database`. Instead, implementation-dependent insertion functionality is achieved by overriding `string_of_query` for `Insert` AST nodes.
  - The `make_insert_returning_query` method is simplified to take an insertion query and a `returning` field, rather than serialised values.