Partially fix #835 (#837)
This patch partially addresses #835, in particular:

```
 query flat { for (x <--factorials) [(a=x.i,b=x.f)]};
 query flat { for (x <- asList(factorials)) [(a=x.i,b=x.f)]};
```

These queries (explicit `flat` modifier) did not run as expected when `shredding` was set to true. The reason for this is that all `Iteration` nodes are implicitly wrapped in a `Query` block if they contain a table generator, and unannotated query blocks took the evaluation policy set by the `shredding` option. Thus:

```
query flat { for (x <--factorials) [(a=x.i,b=x.f)]};
```
was effectively transformed into
```
query flat { query nested { for (x <--factorials) [(a=x.i,b=x.f)]} };
```

causing a mismatch.

This patch fixes this in two ways. Firstly, inner `default` query blocks take the evaluation policy of the outer block, rather than the policy defined by the setting. This should have been the initial design, really.

Secondly, `Query`-block insertion absolutely should not be done by the typechecker. This patch separates it out into a separate pass, `WrapTableIterators`, performed after desugaring.

The other part of #835 is a separate issue, and is out-of-scope for this PR.