Rename option to toggle switch functions.
This commit renames the option `pattern_matching_sugar` to
`switch_functions` as the former is somewhat generic and the latter is
the name of the feature actually being enabled/disabled by toggling
the option.

In addition, this commit also cleans up the test suite for
functions. The non-experimental function features should not be kept
in the same test file as the experimental features -- in particular,
the non-experimental features should not **always** be tested under
the assumption that an experimental flag is enabled.