Distribute MVU JavaScript source files. (#934)
This commit modifies the install script such that the MVU dependencies
`vdom.js` and `virtual-dom.js` are installed along with the MVU
examples.

To make the MVU example work out of the box I have chosen to piggyback
on hardcoded resolution for `jslib.js`. This is a horrible hack. I
also had to change the default configuration file to set the stdlib
path in order for the system to find the MVU Links source files.

I was able to run the MVU todo example application on a fresh
installation of Links with the following command.

```
$ linx $OPAM_SWITCH_PREFIX/share/links/examples/mvu/todomvc/todoMVC.links
```

Resolves #930.