Extensible SQL Serialiser (#941)
# Summary and Motivation

This PR aims to implement #937 by refactoring the monolithic SQL serialiser into an object. As a result, different database drivers may now override parts of the SQL serialiser in order to cleanly and modularly implement driver-specific code. You can see this in action in `pg_database.ml`, which now only needs to override `pp_insert`.

The issue with the previous approach is that it was very difficult to override *parts* of the SQL serialisation code, since it was a monolithic function. The only ways to implement specific functionality for each database was to pass in parameters everywhere (which in turn was an artifact of having the code very tightly-coupled before -- see #832). I got stung badly by this in my concurrent PR, #935 .

# Details

* `sql.ml`: The various printing functions have been made into an object, `Sql.printer`, which takes a `quote` function as its parameter. Methods of `Sql.printer` can be overridden on a per-driver basis, while inheritance allows us to use the default behaviour elsewhere.
* `value.ml`: 
  * The `Value.database` class now takes an `Sql.printer` object as a parameter, which can be specialised per-database. This is now the preferred way to implement driver-specific behaviour, rather than overriding `string_of_query`.
  * `quote_field` has been removed and put inside `Sql.printer`, where it belongs. Code (such as the lens code) can still get access to DB-specific quoting functionality using `db#get_printer#quote_field`.
* Drivers have been updated to take the changes into account.

# Warts

We still have `escape_string` in the `Value.database` class, even though it morally belongs in the `Sql.printer` class. This is because the `pg_database` driver implements `escape_string` by calling through to an underlying `connection` instance which is only defined after initialisation, whereas we need this at initialisation time to pass as a parameter to `pg_printer`.

This patch also does not address the existing wart of `insert returning` statements being tied to `Value.database` rather than in the SQL printer (mainly because of the fact that it can create multiple queries). Future work!