Upgrade Travis image to Ubuntu 20.04 LTS (#956)
Upgrade CI to run postgresql 12 and mysql8. The new package links-mysql8 provides an adapter for connecting to mysql8 databases, whilst the package links-mysql remains compatible with mysql5.7. Note, however, that the package links-mysql is no longer tested by the CI due to difficulties with configuring mysql5.7 on the CI image, instead we test links-mysql8.

Co-authored-by: Daniel Hillerström <daniel.hillerstrom@ed.ac.uk>
Co-authored-by: James Cheney <james.cheney@gmail.com>